Pass in method to reduce duplication
The pilot_submarine and pilot_and_aim_submarine were very similar, the
only difference is how the next coordinate is determined. To reduce
duplication I've added an extra move_submarine method param to the
method, which is applied for each movement.

This allows the identical shell to be reused across both solutions.