Move vectors to closure
Previously I was just cloning the winning vector and returning it, but
that seems inefficient, so I tried to update the closure definition with
lifetimes, so I could just pass back a reference to the right value. I'm
not 100% sure but it seems like that's not really possible, so instead
I'm just moving the two vectors into the closure, and then just
returning the winning one. I think this is still more efficient.