Remove last blank line
Due to VSCode thinking the file was Ruby (for some reason) it was
following up Ruby config and auto inserting a blank line at the end of
the file. This was resulting in the .all method thinking that it was
valid (for some reason) and incorrectly picking the last board as the
correct one.